refactor: move ui package to server/
