+++ server/ui/common/component.go
