+++ server/ui/common/utils.go
